Gems/AWSClientAuth/cdk/requirements.txt
aws-cdk.core
aws-cdk.aws_cognito
aws-cdk.aws_iam